requirements.txt
supabase==2.3.0
python-dotenv==1.0.0
requests==2.32.3
